+++ src/components/CV.jsx
@@ -1,0 +1,9 @@
+const CV = () => {
+  return (
+    <div className="CV">
+      <a href="">Download CV</a>
+      <a href="#contact">Let's Talk</a>
+    </div>
+  )
+}
+export default CV
