+++ src/components/Header.jsx
@@ -1,0 +1,12 @@
+const Header = () => {
+  return (
+    <header>
+      <div className="container header__container">
+        <h5>Hi there, I'm</h5>
+        <h1>Corey Loftus</h1>
+        <h5 className="text-light">Full Stack Developer</h5>
+      </div>
+    </header>
+  )
+}
+export default Header
