+++ src/style/App.scss
@@ -1,26 +1,123 @@
 
 @import url('https://fonts.googleapis.com/css2?family=Chivo+Mono:wght@100;400&family=Noto+Sans:wght@100;400&display=swap');
 
-/* App styling */
-body {
+*{
   margin:0;
-  background-color: black;
-  font-family: 'Noto Sans', sans-serif;
-}
-a:link {
-  color:white;
+  padding:0;
+  border:0;
+  outline:0;
+  box-sizing: border-box;
+  list-style: none;
   text-decoration: none;
-  font-weight:bold;
+}
+
+:root {
+  --color-bg: rgba(0,0,0, .85);
+  --color-bg-variant:#2c2c6c;
+  --color-primary:rgb(3, 189, 189);
+  --color-primary-variant:rgb(214, 105, 245);
+  --color-white:#fff;
+  --color-light:rgba(255,255,255,0.6);
+
+  --transition:all 400ms ease;
+
+  --container-width-lg:75%;
+  --container-width-md:86%;
+  --container-width-sm:90%;
+}
+::-webkit-scrollbar{
+  display:none;
+}
+html{
+  scroll-behavior: smooth;
+}
+body {
+  font-family: 'Noto Sans', sans-serif;
+  background-color: var(--color-bg);
+  color: var(--color-white);
+}
+/* --------Average Screens--------------- */
+
+.container {
+  width: var(--container-width-lg);
+  margin: 0 auto;
+}
+
+h1,
+h2, 
+h3,
+h4,
+h5 {
+  font-weight: 500;
+}
+
+h1 {
+  font-size: 2.5rem;
+}
+
+section {
+  margin-top: 8rem;
+}
+
+section > h2, 
+section > h5 {
+  text-align: center;
+  color: var(--color-light);
+}
+
+section > h2 {
+  color: var(--color-primary);
+  margin-bottom: 3rem;
+}
+
+.text-light {
+  color: var(--color-light);
+
+}
+
+a {
+  color: var(--color-primary);
+  transition: var(--transition);
+}
+
+a:hover {
+  color: var(--color-white);
 }
 a:visited {
-  color:white;
-  text-decoration: none;
+  color:var(--color-white);
+}
+
+.btn {
+  width: max-content;
+  display: inline-block;
+  color: var(--color-primary);
+  padding: 0.75rem 1.2rem;
+  border-radius: 0.4rem;
+  cursor: pointer;
+  border: 1px solid var(--color-primary);
+  transition: var(--transition); 
+}
+
+.btn:hover {
+  background: var(--color-white);
+  color: var(--color-bg);
+  border-color: transparent;
+}
+
+.btn-primary {
+  background: var(--color-primary);
+  color: var(--color-bg);
+}
+
+img {
+  display: block;
+  width: 100%;
+  object-fit: cover;
 }
 
 .App {
   height:100%;
-  color:white;
-  background-color: rgba(85, 85, 85, 0.25);
+
   ul>li {
     list-style: none;
   }
@@ -59,8 +156,8 @@ a:visited {
   }
   .nav-container a:hover {
     background-color: #dbdbdb;
-    color:#BB2649;
-    transition: ease-in-out .2s;
+    color:var(--color-primary-variant);
+    transition: var(--transition);
   }
   .links-box {
     display:flex;
@@ -101,6 +198,59 @@ a:visited {
   .project-desc{
     margin:1rem;
   }
+  .portfolio__container {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    gap: 2rem;
+}
+
+.portfolio__item {
+    background: var(--color-bg-variant);
+    padding: 1.5rem;
+    border-radius: 2rem;
+    border: 1px solid transparent;
+    transition: var(--transition);
+}
+
+.portfolio__item:hover {
+    background: transparent;
+    border-color: var(--color-primary-variant);
+    cursor: default;
+}
+
+.portfolio__item-image {
+    border-radius: 1rem;
+    overflow: hidden;
+}
+
+.portfolio__item h3 {
+    margin: 1.2rem 0 2rem;
+}
+
+.portfolio__item-cta {
+    display: flex;
+    gap: 1rem;
+    margin-bottom: 1rem;
+}
+
+/* -----medium screen----- */
+
+@media screen and (max-width: 1024px){
+    .portfolio__container{
+    gap: 1.2rem;
+    }
+
+}
+
+/* -----small screen----- */
+
+@media screen and (max-width: 600px){
+    .portfolio__container{
+        grid-template-columns: 1fr;
+        gap: 1rem;
+    }
+
+}
 }
 
 .About {
@@ -113,7 +263,7 @@ a:visited {
     justify-content: space-evenly;
   }
   .color-box {
-    background-color: gray;
+    background-color: var(--color-light);
     border-radius: .4rem;
     padding:.2rem;
     margin:.8rem;
